dubbo-samples-merge:change groupId and package to org.apache.dubbo
