+++ dubbo-samples-merge/src/main/java/org/apache/dubbo/samples/merge/EmbeddedZooKeeper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.samples.merge;
+package org.apache.dubbo.samples.merge;
 
 import java.io.File;
 import java.lang.reflect.Method;
