+++ dubbo-samples-merge/src/main/java/org/apache/dubbo/samples/merge/MergeConsumer2.java
@@ -17,21 +17,21 @@
  *
  */
 
-package com.alibaba.dubbo.samples.merge;
+package org.apache.dubbo.samples.merge;
 
 import java.util.List;
 
-import com.alibaba.dubbo.samples.merge.api.MergeService;
+import org.apache.dubbo.samples.merge.api.MergeService;
 
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 /**
- * MergeConsumer
+ * MergeConsumer2
  */
-public class MergeConsumer {
+public class MergeConsumer2 {
 
     public static void main(String[] args) throws Exception {
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/merge-consumer.xml"});
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/merge-consumer2.xml"});
         context.start();
         MergeService mergeService = (MergeService) context.getBean("mergeService");
         for (int i = 0; i < Integer.MAX_VALUE; i++) {
