+++ dubbo-samples-merge/src/main/java/org/apache/dubbo/samples/merge/api/MergeService.java
@@ -17,23 +17,15 @@
  *
  */
 
-package com.alibaba.dubbo.samples.merge.impl;
+package org.apache.dubbo.samples.merge.api;
 
-import java.util.ArrayList;
 import java.util.List;
 
-import com.alibaba.dubbo.samples.merge.api.MergeService;
-
 /**
- * MenuServiceImpl
+ * MergeService
  */
-public class MergeServiceImpl2 implements MergeService {
+public interface MergeService {
 
-    public List<String> mergeResult() {
-        List<String> menus = new ArrayList<String>();
-        menus.add("group-2.1");
-        menus.add("group-2.2");
-        return menus;
-    }
+    List<String> mergeResult();
 
 }
