+++ dubbo-samples-merge/src/main/java/org/apache/dubbo/samples/merge/impl/MergeServiceImpl3.java
@@ -17,18 +17,19 @@
  *
  */
 
-package com.alibaba.dubbo.samples.merge.impl;
+package org.apache.dubbo.samples.merge.impl;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import com.alibaba.dubbo.samples.merge.api.MergeService;
+import org.apache.dubbo.samples.merge.api.MergeService;
 
 /**
  * MenuServiceImpl
  */
 public class MergeServiceImpl3 implements MergeService {
 
+    @Override
     public List<String> mergeResult() {
         List<String> menus = new ArrayList<String>();
         menus.add("group-3.1");
