+++ dubbo-samples-merge/src/main/resources/spring/merge-consumer2.xml
@@ -28,7 +28,7 @@
 
     <dubbo:registry address="zookeeper://127.0.0.1:2181"/>
 
-    <dubbo:reference id="mergeService" interface="com.alibaba.dubbo.samples.merge.api.MergeService"
+    <dubbo:reference id="mergeService" interface="org.apache.dubbo.samples.merge.api.MergeService"
                      group="merge2,merge3"/>
 
 </beans>
