+++ dubbo-samples-merge/src/main/resources/spring/merge-provider2.xml
@@ -30,8 +30,8 @@
 
     <dubbo:protocol name="dubbo" port="20881"/>
 
-    <bean id="mergeService3" class="com.alibaba.dubbo.samples.merge.impl.MergeServiceImpl3"/>
+    <bean id="mergeService3" class="org.apache.dubbo.samples.merge.impl.MergeServiceImpl3"/>
 
-    <dubbo:service group="merge3" interface="com.alibaba.dubbo.samples.merge.api.MergeService" ref="mergeService3"/>
+    <dubbo:service group="merge3" interface="org.apache.dubbo.samples.merge.api.MergeService" ref="mergeService3"/>
 
 </beans>
